refactor(backup): own launch tile + route, remove heavy backup settings page
Backup gets the same treatment disk-import got in #1953: its own top-level
launch tile + route, and the bulky backup UI leaves the Settings IA entirely
(rip-and-replace, no compat shim).

- Move the ~1700-line backup/restore page + its _lib (useBackupState,
  LocalTargetPicker, ExternalBackupDestinationSection) out of
  settings/backups into a new (dashboard)/backup app with its own
  PageHeader-wrapped layout; add a "Backup & restore" launch tile in
  navigation.ts mirroring the disk-import tile.
- Drop the SettingsContext coupling: the standalone app loads nodes via the
  getNodes server action instead of useSettings().
- Carve the backup-specific helpers into the new app's _lib/helpers.ts and
  remove them from the settings helpers.
- Remove the Backups group from the settings IA (ia.ts) so the settings
  shell/search/3-tier no longer reference it; the actions still run in the
  capped backup worker (#1955).

Closes #1958
